исправлены критические ошибки в AuthPage и RegPage
