+++ resources/frontend/src/Components/RegPage/module.regPage.css
@@ -1,0 +1,25 @@
+.wrapper {
+    width: 400px;
+    display: block;
+    margin: 0 auto;
+    margin-top: 200px;
+  }
+  
+  .form-login {
+    display: flex;
+    justify-content: space-between;
+  }
+  .form {
+    display: flex;
+    flex-direction: column;
+  }
+  .btn {
+    margin-top: 25px;
+  }
+  small{
+    font-size: 10px;
+  }
+  h3 {
+    text-align: center;
+    padding-bottom: 15px;
+  }
